final v4 Moulinette passed

+++ ft_strlen.c
@@ -6,15 +6,16 @@
 /*   By: pgorner <pgorner@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/12 10:04:09 by pgorner           #+#    #+#             */
-/*   Updated: 2022/10/18 13:04:27 by pgorner          ###   ########.fr       */
+/*   Updated: 2022/11/02 15:04:22 by pgorner          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 /* FUNCTION EXPLANATION: 
 ** Function outputs string length
 */
+#include "libft.h"
 
-unsigned int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	i;
